requirements.txt
fastapi~=0.111.0
uvicorn
sqlalchemy~=2.0.30
alembic
pydantic~=2.7.1
python-dotenv~=1.0.1
passlib~=1.7.4
bcrypt
pytelegrambotapi
stripe
telebot